Extract signature_api.rs
