+++ rust/rubydex-sys/src/lib.rs
@@ -76,4 +76,5 @@ pub mod graph_api;
 pub mod location_api;
 pub mod name_api;
 pub mod reference_api;
+pub mod signature_api;
 pub mod utils;
